+++ cmsplugin_phlog/forms/__init__.py
@@ -1,0 +1,1 @@
+from cmsplugin_phlog.forms.forms import GalleryPluginForm