+++ cmsplugin_phlog/media/js/init.js
@@ -1,0 +1,5 @@
+(function ($){
+	$(document).ready(function(){
+	    $.fn.cmsPatchCSRF();
+	});
+})(jQuery);