+++ cmsplugin_phlog/migrations/0002_v0_2.py
@@ -1,0 +1,152 @@
+# encoding: utf-8
+import datetime
+from south.db import db
+from south.v2 import SchemaMigration
+from django.db import models
+
+class Migration(SchemaMigration):
+
+    def forwards(self, orm):
+        
+        # Adding model 'GalleryPlugin'
+        db.create_table('cmsplugin_phlog_gallery', (
+            ('cmsplugin_ptr', self.gf('django.db.models.fields.related.OneToOneField')(to=orm['cms.CMSPlugin'], unique=True, primary_key=True)),
+            ('template', self.gf('django.db.models.fields.CharField')(max_length=255, blank=True)),
+        ))
+        db.send_create_signal('cmsplugin_phlog', ['GalleryPlugin'])
+
+        # Adding model 'PhlogPhotoPlugin'
+        db.create_table('cmsplugin_phlog_photo', (
+            ('cmsplugin_ptr', self.gf('django.db.models.fields.related.OneToOneField')(to=orm['cms.CMSPlugin'], unique=True, primary_key=True)),
+        ))
+        db.send_create_signal('cmsplugin_phlog', ['PhlogPhotoPlugin'])
+
+        # Adding model 'PhlogPhoto'
+        db.create_table('phlog_photo', (
+            ('id', self.gf('django.db.models.fields.AutoField')(primary_key=True)),
+            ('image', self.gf('django.db.models.fields.files.ImageField')(max_length=100)),
+            ('date_taken', self.gf('django.db.models.fields.DateTimeField')(null=True, blank=True)),
+            ('view_count', self.gf('django.db.models.fields.PositiveIntegerField')(default=0)),
+            ('crop_from', self.gf('django.db.models.fields.CharField')(default='center', max_length=10, blank=True)),
+            ('effect', self.gf('django.db.models.fields.related.ForeignKey')(blank=True, related_name='phlogphoto_related', null=True, to=orm['photologue.PhotoEffect'])),
+            ('title', self.gf('django.db.models.fields.CharField')(max_length=255)),
+            ('caption', self.gf('django.db.models.fields.TextField')(null=True, blank=True)),
+            ('plugin', self.gf('django.db.models.fields.related.ForeignKey')(related_name='photos', to=orm['cms.CMSPlugin'])),
+        ))
+        db.send_create_signal('cmsplugin_phlog', ['PhlogPhoto'])
+
+
+    def backwards(self, orm):
+        
+        # Deleting model 'GalleryPlugin'
+        db.delete_table('cmsplugin_phlog_gallery')
+
+        # Deleting model 'PhlogPhotoPlugin'
+        db.delete_table('cmsplugin_phlog_photo')
+
+        # Deleting model 'PhlogPhoto'
+        db.delete_table('phlog_photo')
+
+
+    models = {
+        'cms.cmsplugin': {
+            'Meta': {'object_name': 'CMSPlugin'},
+            'creation_date': ('django.db.models.fields.DateTimeField', [], {'default': 'datetime.datetime.now'}),
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'language': ('django.db.models.fields.CharField', [], {'max_length': '15', 'db_index': 'True'}),
+            'level': ('django.db.models.fields.PositiveIntegerField', [], {'db_index': 'True'}),
+            'lft': ('django.db.models.fields.PositiveIntegerField', [], {'db_index': 'True'}),
+            'parent': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['cms.CMSPlugin']", 'null': 'True', 'blank': 'True'}),
+            'placeholder': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['cms.Placeholder']", 'null': 'True'}),
+            'plugin_type': ('django.db.models.fields.CharField', [], {'max_length': '50', 'db_index': 'True'}),
+            'position': ('django.db.models.fields.PositiveSmallIntegerField', [], {'null': 'True', 'blank': 'True'}),
+            'rght': ('django.db.models.fields.PositiveIntegerField', [], {'db_index': 'True'}),
+            'tree_id': ('django.db.models.fields.PositiveIntegerField', [], {'db_index': 'True'})
+        },
+        'cms.placeholder': {
+            'Meta': {'object_name': 'Placeholder'},
+            'default_width': ('django.db.models.fields.PositiveSmallIntegerField', [], {'null': 'True'}),
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'slot': ('django.db.models.fields.CharField', [], {'max_length': '50', 'db_index': 'True'})
+        },
+        'cmsplugin_phlog.galleryplugin': {
+            'Meta': {'object_name': 'GalleryPlugin', 'db_table': "'cmsplugin_phlog_gallery'", '_ormbases': ['cms.CMSPlugin']},
+            'cmsplugin_ptr': ('django.db.models.fields.related.OneToOneField', [], {'to': "orm['cms.CMSPlugin']", 'unique': 'True', 'primary_key': 'True'}),
+            'template': ('django.db.models.fields.CharField', [], {'max_length': '255', 'blank': 'True'})
+        },
+        'cmsplugin_phlog.phloggalleryplugin': {
+            'Meta': {'object_name': 'PhlogGalleryPlugin', 'db_table': "'cmsplugin_phloggalleryplugin'", '_ormbases': ['cms.CMSPlugin']},
+            'cmsplugin_ptr': ('django.db.models.fields.related.OneToOneField', [], {'to': "orm['cms.CMSPlugin']", 'unique': 'True', 'primary_key': 'True'}),
+            'gallery': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['photologue.OrderedGallery']"}),
+            'limit': ('django.db.models.fields.PositiveIntegerField', [], {'default': '0'}),
+            'order': ('django.db.models.fields.CharField', [], {'default': "'gallery'", 'max_length': '32'}),
+            'public': ('django.db.models.fields.BooleanField', [], {'default': 'True'}),
+            'template': ('django.db.models.fields.CharField', [], {'max_length': '255', 'blank': 'True'})
+        },
+        'cmsplugin_phlog.phlogphoto': {
+            'Meta': {'object_name': 'PhlogPhoto', 'db_table': "'phlog_photo'"},
+            'caption': ('django.db.models.fields.TextField', [], {'null': 'True', 'blank': 'True'}),
+            'crop_from': ('django.db.models.fields.CharField', [], {'default': "'center'", 'max_length': '10', 'blank': 'True'}),
+            'date_taken': ('django.db.models.fields.DateTimeField', [], {'null': 'True', 'blank': 'True'}),
+            'effect': ('django.db.models.fields.related.ForeignKey', [], {'blank': 'True', 'related_name': "'phlogphoto_related'", 'null': 'True', 'to': "orm['photologue.PhotoEffect']"}),
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'image': ('django.db.models.fields.files.ImageField', [], {'max_length': '100'}),
+            'plugin': ('django.db.models.fields.related.ForeignKey', [], {'related_name': "'photos'", 'to': "orm['cms.CMSPlugin']"}),
+            'title': ('django.db.models.fields.CharField', [], {'max_length': '255'}),
+            'view_count': ('django.db.models.fields.PositiveIntegerField', [], {'default': '0'})
+        },
+        'cmsplugin_phlog.phlogphotoplugin': {
+            'Meta': {'object_name': 'PhlogPhotoPlugin', 'db_table': "'cmsplugin_phlog_photo'", '_ormbases': ['cms.CMSPlugin']},
+            'cmsplugin_ptr': ('django.db.models.fields.related.OneToOneField', [], {'to': "orm['cms.CMSPlugin']", 'unique': 'True', 'primary_key': 'True'})
+        },
+        'photologue.orderedgallery': {
+            'Meta': {'ordering': "['-date_added']", 'object_name': 'OrderedGallery'},
+            'date_added': ('django.db.models.fields.DateTimeField', [], {'default': 'datetime.datetime.now'}),
+            'description': ('django.db.models.fields.TextField', [], {'blank': 'True'}),
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'is_public': ('django.db.models.fields.BooleanField', [], {'default': 'True'}),
+            'photos': ('django.db.models.fields.related.ManyToManyField', [], {'related_name': "'ordered_galleries'", 'to': "orm['photologue.Photo']", 'through': "orm['photologue.PhotoOrdering']", 'blank': 'True', 'symmetrical': 'False', 'null': 'True'}),
+            'tags': ('tagging.fields.TagField', [], {}),
+            'title': ('django.db.models.fields.CharField', [], {'unique': 'True', 'max_length': '100'}),
+            'title_slug': ('django.db.models.fields.SlugField', [], {'unique': 'True', 'max_length': '50', 'db_index': 'True'})
+        },
+        'photologue.photo': {
+            'Meta': {'ordering': "['-date_added']", 'object_name': 'Photo'},
+            'caption': ('django.db.models.fields.TextField', [], {'blank': 'True'}),
+            'crop_from': ('django.db.models.fields.CharField', [], {'default': "'center'", 'max_length': '10', 'blank': 'True'}),
+            'date_added': ('django.db.models.fields.DateTimeField', [], {'default': 'datetime.datetime.now'}),
+            'date_taken': ('django.db.models.fields.DateTimeField', [], {'null': 'True', 'blank': 'True'}),
+            'effect': ('django.db.models.fields.related.ForeignKey', [], {'blank': 'True', 'related_name': "'photo_related'", 'null': 'True', 'to': "orm['photologue.PhotoEffect']"}),
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'image': ('django.db.models.fields.files.ImageField', [], {'max_length': '100'}),
+            'is_public': ('django.db.models.fields.BooleanField', [], {'default': 'True'}),
+            'tags': ('tagging.fields.TagField', [], {}),
+            'title': ('django.db.models.fields.CharField', [], {'unique': 'True', 'max_length': '100'}),
+            'title_slug': ('django.db.models.fields.SlugField', [], {'unique': 'True', 'max_length': '50', 'db_index': 'True'}),
+            'view_count': ('django.db.models.fields.PositiveIntegerField', [], {'default': '0'})
+        },
+        'photologue.photoeffect': {
+            'Meta': {'object_name': 'PhotoEffect'},
+            'background_color': ('django.db.models.fields.CharField', [], {'default': "'#FFFFFF'", 'max_length': '7'}),
+            'brightness': ('django.db.models.fields.FloatField', [], {'default': '1.0'}),
+            'color': ('django.db.models.fields.FloatField', [], {'default': '1.0'}),
+            'contrast': ('django.db.models.fields.FloatField', [], {'default': '1.0'}),
+            'description': ('django.db.models.fields.TextField', [], {'blank': 'True'}),
+            'filters': ('django.db.models.fields.CharField', [], {'max_length': '200', 'blank': 'True'}),
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'name': ('django.db.models.fields.CharField', [], {'unique': 'True', 'max_length': '30'}),
+            'reflection_size': ('django.db.models.fields.FloatField', [], {'default': '0'}),
+            'reflection_strength': ('django.db.models.fields.FloatField', [], {'default': '0.59999999999999998'}),
+            'sharpness': ('django.db.models.fields.FloatField', [], {'default': '1.0'}),
+            'transpose_method': ('django.db.models.fields.CharField', [], {'max_length': '15', 'blank': 'True'})
+        },
+        'photologue.photoordering': {
+            'Meta': {'object_name': 'PhotoOrdering'},
+            'gallery': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['photologue.OrderedGallery']"}),
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'order': ('django.db.models.fields.PositiveIntegerField', [], {'default': '9999'}),
+            'photo': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['photologue.Photo']"})
+        }
+    }
+
+    complete_apps = ['cmsplugin_phlog']