+++ HISTORY.rst
@@ -1,3 +1,13 @@
+0.2 (2011/08/04)
+----------------
+
+* Deprecated 0.1 gallery
+* Added a gallery plugin that allows for other plugins to be added as items in 
+  the gallery
+* Added a Photo plugin for adding photologue photos to a page (and not the site 
+  wide photo manager)
+* Added South integration
+
 0.1 (2011/05/11)
 ----------------